Move GCM code out of reference routes
