+++ conf/routes
@@ -2,16 +2,6 @@
 # This file defines all application routes (Higher priority routes first)
 # ~~~~
 
-#==================
-# REFERENCE ROUTES
-#==================
-# Database query reference
-GET     /ref/dbquery                controllers.Reference.dbQuery
-
-# GCM setup reference
-POST    /ref/gcmtoken              controllers.Reference.gcmToken
-POST    /ref/gcmsend               controllers.Reference.gcmNotify
-
 #================
 # WHEREU@ ROUTES
 #================
